Gate ARM64-accelerated impl on hwcap() having the HWCAP_PMULL flag.
The hardware-accelerated CRC32C implementation that takes advantage of
ARM64 instructions is currently runtime-gated on hwcap() returning a
value that has the HWCAP_CRC32 flag set. This covers the
__crc32c{b,h,w,d} intrinsics, but does not cover the vmull_p64 call. The
later should be gated on the presence of the HWCAP_PMULL flag.

This is a speculative fix for Chrome crashes observed at the first
vmull_64 callsite on MSM8916-based boards.