Migrate to 0/1 feature detection macros. (#10)
The build system previously defined macros corresponding to detected
features, and left the other macros undefined. After this change, all
the feature detection macros are defined. Macros for detected features
are 1, and the other macros are 0.

This makes it slightly more difficult to embed crc32c in a system that
doesn't use the CMake generator, but reduces the possibility of the
library not using features due to accidental misconfiguration.