Add Bazel support.
This gets the code to compile and the tests to pass with Bazel. To
ease the maintenance burden the list of sources files is automatically
generated from the CMakeList.txt file, so developers need to update the
list in one place.

There are probably many optimizations that are not enabled by this change,
but at least the code compiles and the test pass. We can iterate on the
optimization front later.